test codecov config file
