+++ test/ui/drawer_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:image_test_utils/image_test_utils.dart';
 import 'package:maybank2u/application/injector.dart';
-import 'package:maybank2u/main.dart';
+import 'package:maybank2u/application/application.dart';
 import 'package:maybank2u/ui/intro/widgets/drawer/drawer.dart';
 
 void main() {
